updated mongo collections and routes
