+++ config/mongoCollections.js
@@ -1,7 +1,5 @@
 const dbConnection = require('./mongoConnection');
 
-/* This will allow you to have one reference to each collection per app */
-/* Feel free to copy and paste this this */
 const getCollectionFn = (collection) => {
   let _col = undefined;
 
@@ -15,9 +13,8 @@ const getCollectionFn = (collection) => {
   };
 };
 
-/* Now, you can list your collections here: */
 module.exports = {
-  //get closet collection
-  closets: getCollectionFn('closets'),
-  // albums: getCollectionFn('albums')
+  users: getCollectionFn('users'),
+  clothes: getCollectionFn('clothes'),
+  outfits: getCollectionFn('outfits'),
 };
