+++ views/layouts/main.handlebars
@@ -10,7 +10,7 @@
         <div class="topnav">
             <a class="active" href="/">Home</a>
             {{#if signed_in}}
-            <a href="/">My Clothes</a>
+            <a href="/clothes">My Clothes</a>
             <a href="/">My Outfits</a>
             <a href="/">Statistics</a>
             <a href="/">Profile</a>
